# Imporvement of performance : Reduce the Java running count