# What is it
- #6465 is the first reqeust.  So I followed up the albert advice.
	- in case of debug you use printf statements, I think that here there better options would be
		- create an -d flag (e.g. -d plantuml) that gives the information, so no need for using the configuration option
		- Follow up : create -d plantuml [debug.h, debug.cpp]
	- usage of printf
		- printf should not be used,all output is steered over calls to routines as defined in message.cpp,
		- Follow up : use Debug::Print(Debug::Plantuml, ...    and msg()
	- usage of std::string / std::map
		- in doxygen the string manipulation is. mostly, done by means of Qt classes (e.g. classes QCString, QString). For consistency these classes should be used.
		- map manipulation is also done by means of Qt classes (dictionary and list classes)
		- Follow up : use QDict QList QCString [plantuml.h, plantuml.cpp]
	- dirent / opendir /readdir
		- when browsing through directories the routines line readFileOrDirectory should probably be used or the techniques used in this routine
		- Follow up : I will follow when I try to reduce redundancy of creating plantuml image.

- Keep the seperate pu files  because of easy debugging.

- Run java minimally
	- Test Case : 4 plantuml
	- Original Run without PLANTUML_RUN_JAVA_ONCE
		- 8.2 sec = 4 times * ( 1.6(java vm) + 0.1 * 1(load multiple file) + process each plantuml (0.35) *1 ) <- prediction
	- New with PLANTUML_RUN_JAVA_ONCE
		- 3.499 sec = 1.6(java vm) + 0.1 * 4(load multiple file) + process each plantuml (0.35) *4 <- prediction
	- Creating Java Virtual Machine has a big portion.

- Result
	- Improving Performance : 8.2 seconds -> 3.499 seconds  (in case of 4 plantuml)
	- Performance in theory  :  1 to 4 times faster 
 

# Configuration
- add configuration value in Doxyfile
	- PLANTUML_RUN_JAVA_ONCE = YES

# Debugging
- doxygen -d plantuml
- doxygen -d time -d plantuml

# Tested with following options
- source code of 3 plantuml
- PLANTUML_RUN_JAVA_ONCE = YES  or    NO
- DOT_IMAGE_FORMAT       = png  or    svg
- DOT_CLEANUP            = YES  or    NO
- -d plantuml  -d time