@albert-github 2018-08-27
I ran a simple test too, 40 pu files (distributed over 2 directories), and went from 44 to 18 sec. so also a considerable improvement.

In the debug statements the `__PRETTY_FUNCTION__` is a GCC extension (see e.g. https://stackoverflow.com/questions/4384765/whats-the-difference-between-pretty-function-function-func) and should not be used. On native Windows (I use the 2013 version i.e. 18.00.31101) and the `__PRETTY_FUNCTION__` is unknown. I also had problems with `__FUNC__` and `__func__` so best is just not to use it or use `__FILE__` the debug will give enough information in that way too.

- In the `PLANTUML_RUN_JAVA_ONCE ` all the `.pu` files are given to the `plantuml.jar` with their full path (see `-d extcmd` when running). I think this will lead to problems when there are quite a few files ("command line" overflow). Seen this merging the files into one file like done for formulas. A small test with one file revealed that it didn't bring a lot / nothing time wise, but here the problem with the "command line" overflow shouldn't / can't happen.
- I think it still would make sense to implement the 'md5' possibility for, especially, the non `PLANTUML_RUN_JAVA_ONCE ` version, i.e. if the file already exists don't try to recreate the image.